requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
accelerate>=0.25.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.66.0
decord>=0.6.0

